# Conflicts:
#	src/main/java/com/gedi/projectmanagement/base/GeneralMapper.java
#	src/main/java/com/gedi/projectmanagement/controller/SysUserController.java
#	src/main/java/com/gedi/projectmanagement/controller/WeekReportController.java
#	src/main/java/com/gedi/projectmanagement/dao/system/SysUserMapper.java
#	src/main/java/com/gedi/projectmanagement/model/system/SysUser.java
#	src/main/java/com/gedi/projectmanagement/scheduled/DepartSynchScheduled.java
#	src/main/java/com/gedi/projectmanagement/scheduled/UserSynchScheduled.java
#	src/main/java/com/gedi/projectmanagement/service/impl/system/SysDepartmentServiceImpl.java
#	src/main/java/com/gedi/projectmanagement/service/impl/system/SysUserServiceImpl.java
#	src/main/java/com/gedi/projectmanagement/service/system/SysUserService.java
#	src/main/resources/mapper/SysUserMapper.xml
#	src/test/java/com/gedi/projectmanagement/SysUserTests.java

+++ src/main/java/com/gedi/projectmanagement/service/impl/ProjectPlanserviceImpl.java
@@ -1,19 +1,23 @@
 package com.gedi.projectmanagement.service.impl;
 
-import com.gedi.projectmanagement.dao.ActionItemMapper;
-import com.gedi.projectmanagement.dao.ProjectPlanMapper;
-import com.gedi.projectmanagement.dao.ProjectUserMediumMapper;
+import com.gedi.projectmanagement.dao.*;
+import com.gedi.projectmanagement.model.Journal;
 import com.gedi.projectmanagement.model.ProjectPlan;
 import com.gedi.projectmanagement.model.ProjectPlanList;
+import com.gedi.projectmanagement.model.Weekreport;
 import com.gedi.projectmanagement.service.ProjectPlanService;
 import com.gedi.projectmanagement.util.UUIDUtil;
 import com.gedi.projectmanagement.vo.CodeAndMsg;
 import com.gedi.projectmanagement.vo.CodeAndMsgUtil;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.Resource;
+import javax.servlet.http.HttpServletRequest;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 /**
  * @author wangyanbin
@@ -32,25 +36,34 @@ public class ProjectPlanserviceImpl implements ProjectPlanService {
     @Resource
     private ProjectUserMediumMapper projectUserMediumMapper;
 
+    @Autowired
+    private WeekreportMapper weekreportMapper;
+
+    @Autowired
+    private JournalMapper journalMapper;
+
     //查询所有用于列表展示
     @Override
-    public CodeAndMsg selectById() {
+    public List<ProjectPlan> selectById() {
         List<ProjectPlan> projectPlans = projectPlanMapper.selectById();
-        if (projectPlans != null) {
-            return CodeAndMsgUtil.setOK("查询成功", projectPlans);
+        if (projectPlans != null && projectPlans.size() > 0) {
+            return projectPlans;
         } else {
-            return CodeAndMsgUtil.setERROR("查询失败", null);
+            return null;
         }
     }
 
-
     //新增项目总计划
     @Override
-    public String addProject(ProjectPlan projectPlan) {
+    public String addProject(ProjectPlan projectPlan, HttpServletRequest request) {
         projectPlan.setpId(UUIDUtil.getUUID2());
         String startTime = projectPlan.getpStartTime();
         String annualTime = startTime.substring(0, 4);
         projectPlan.setAnnualTime(annualTime);
+        /*HttpSession session = request.getSession();
+        User user = (User) session.getAttribute("user");
+        projectPlan.setCreater(user.getuName());*/
+        projectPlan.setCreater("张帅");
         projectPlanMapper.addProject(projectPlan);
         return "success";
     }
@@ -108,12 +121,48 @@ public class ProjectPlanserviceImpl implements ProjectPlanService {
      */
     @Override
     public String deleteProjectBypId(String pId) {
-        int i = projectPlanMapper.deleteProjectBypId(pId);
-        if (i > 0) {
-            int j = actionItemMapper.deleteActionItemBypId(pId);
-            int k = projectUserMediumMapper.deletePUMBypId(pId);
+        List<Weekreport> weekreports = weekreportMapper.selectBypId(pId);
+        List<Journal> journals = journalMapper.selectBypId(pId);
+        if ((weekreports != null && weekreports.size() > 0) || (journals != null && journals.size() > 0)) {
+            return "error";
+        } else {
+            int i = projectPlanMapper.deleteProjectBypId(pId);
+            if (i > 0) {
+                int j = actionItemMapper.deleteActionItemBypId(pId);
+                int k = projectUserMediumMapper.deletePUMBypId(pId);
+            }
+            return "success";
         }
-        return "success";
+    }
+
+    /**
+     * 根据pId修改项目总计划
+     *
+     * @param projectPlan
+     * @return
+     */
+    @Override
+    public Map updateProjectBypId(ProjectPlan projectPlan) {
+        Map map = new HashMap();
+        String startTime = projectPlan.getpStartTime();
+        if (startTime != null && startTime != "") {
+            String annualTime = startTime.substring(0, 4);
+            projectPlan.setAnnualTime(annualTime);
+            int i = projectPlanMapper.updateProjectBypId(projectPlan);
+            if (i > 0) {
+                map.put("success", "success");
+            } else {
+                map.put("error", "error");
+            }
+        } else {
+            int i = projectPlanMapper.updateProjectBypId(projectPlan);
+            if (i > 0) {
+                map.put("success", "success");
+            } else {
+                map.put("error", "error");
+            }
+        }
+        return map;
     }
 
 }
